Moving app config to a config js file and fixing the da push url to prioritise contentUrl
